code/requirements.txt
nibabel==5.2.1
numpy==2.0.1
scipy==1.14.0
torchio==0.19.9
scikit-image==0.24.0
tensorboard==2.17.0
tqdm